+++ src/main/java/cfbastian/renderer3d/math/Vector3.java
@@ -1,16 +1,16 @@
 package cfbastian.renderer3d.math;
 
 public class Vector3 {
-    public double x, y, z;
+    public float x, y, z;
 
-    public Vector3(double x, double y, double z) {
+    public Vector3(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public double[] toArray()
+    public float[] toArray()
     {
-        return new double[]{x, y, z};
+        return new float[]{x, y, z};
     }
 }
